support for windows

+++ src/kernels/cpu/concat.cc
@@ -33,7 +33,7 @@ template <typename T> class NaiveConcat : public CpuKernelWithoutConfig {
             auto inPtr = input->getRawDataPtr<T *>(),
                  outPtr = output->getRawDataPtr<T *>();
 #pragma omp parallel for
-            for (size_t iOffset = 0; iOffset < inSize; ++iOffset) {
+            for (long long iOffset = 0; iOffset < inSize; ++iOffset) {
                 auto oOffset = iOffset % localBlockOffset + innerOffset +
                                iOffset / localBlockOffset * blockOffset;
                 // output->setData(oOffset, input->getData(iOffset));
